code organization improved
